# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Miva Campus Maintenance API",
    "description": "RESTful API for the Miva Open University Campus Maintenance Complaint & Service Request Platform. Provides CRUD operations for service requests, user management, task assignment, status logging, and data export.",
    "version": "1.0.0",
    "contact": {
      "name": "Miva Open University — Facilities Management",
      "email": "facilities@miva.edu.ng"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://mivaca.vercel.app/api",
      "description": "Production Server (Vercel)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token obtained from auth session"
      }
    },
    "schemas": {
...
      "AssignRequestBody": {
        "type": "object",
        "required": ["officer_id"],
        "properties": {
          "officer_id": { "type": "string", "format": "uuid" },
          "notes": { "type": "string", "maxLength": 500 }
        }
      },
...
    }
  },
  "security": [{ "BearerAuth": [] }],
  "paths": {
...
    "/requests/{id}": {
      "get": {
        "tags": ["Service Requests"],
        "summary": "Get a single request with full detail",
        "description": "Returns a service request with joined category, requester profile, assignment, and full status log (audit trail).",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Success" },
          "403": { "description": "Access denied — not the owner, assigned officer, or admin" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Service Requests"],
        "summary": "Update a request (Officer/Admin only)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" }, "priority": { "type": "string" } } } } }
        },
        "responses": {
          "200": { "description": "Updated" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "tags": ["Service Requests"],
        "summary": "Cancel a request (Admin only)",
        "description": "Soft-deletes the request by setting its status to 'cancelled'.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Cancelled" },
          "403": { "description": "Forbidden — Admin only" }
        }
      }
    },
    "/requests/{id}/assign": {
      "post": {
        "tags": ["Assignments"],
        "summary": "Assign a maintenance officer to a request (Admin only)",
        "description": "Assigns the request to an officer, updates status to 'assigned', logs the action, and triggers an email alert to the officer.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AssignRequestBody" } } }
        },
        "responses": {
          "201": { "description": "Assigned successfully" },
          "403": { "description": "Forbidden — Admin only" }
        }
      }
    },
...
    "/export": {
      "get": {
        "tags": ["Export"],
        "summary": "Export requests to CSV (Admin only)",
        "description": "Generates and streams a CSV file of all service requests with optional filters.",
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": { "description": "CSV file download", "content": { "text/csv": { "schema": { "type": "string" } } } },
          "403": { "description": "Forbidden — Admin only" }
        }
      }
    },
...
    "/requests/{id}/comments": {
      "get": {
        "tags": ["Comments"],
        "summary": "List comments for a request",
        "description": "Fetches all comments/messages for a specific request. Internal comments are only visible to officers and admins.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "List of comments" } }
      },
      "post": {
        "tags": ["Comments"],
        "summary": "Add a comment",
        "description": "Adds a new message to the request. Supports internal notes and file attachments.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "content": { "type": "string" }, "is_internal": { "type": "boolean" }, "attachments": { "type": "array", "items": { "type": "string" } } } } } }
        },
        "responses": { "201": { "description": "Comment added" } }
      }
    },
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "List request categories",
        "responses": { "200": { "description": "List of categories" } }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create category (Admin only)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } } } } },
        "responses": { "201": { "description": "Category created" } }
      }
    },
...
  }
}